refactor: externalize PaymentConfirmModal
