+++ src/components/Send.module.css
@@ -154,13 +154,6 @@ input[type='number'] {
   filter: blur(2px);
 }
 
-.infoIcon {
-  margin: 2px 0 0 0.25rem;
-  color: var(--bs-gray-500);
-  border: 1px solid var(--bs-gray-500);
-  border-radius: 50%;
-}
-
 .accordion-button {
   border: none;
   width: 100%;
